requirements.txt
ipykernel
jupyter
python-dotenv

langchain
langsmith
langchain_community
langchain-classic
langserve

langchain-openrouter
langchain_openai
langchain-ollama
langchain-groq
langchain_huggingface

langchain-text-splitters
sentence_transformers

langchain_chroma
faiss-cpu
chromadb

pypdf
pyowm
pymupdf
arxiv
wikipedia
beautifulsoup4

streamlit
uvicorn
fastapi
sse_starlette